Add classes defined in Clojure's core library defined via deftype For example, class clojure.core.Vec is defined via deftype.
Use class names with packages for classes and interfaces outside of
the clojure.lang package, but with abbreviations for many of them that
are shown in the legend, e.g. "j.l" for "java.lang".

Change the shape around Java interfaces from a diamond to a
rectangular box, so they take up less horizontal width for the new
longer names.

Add to the legend some notes on the badges.